Huge refactorisation

+++ public-transport/src/main/java/ija/proj/Tram.java
@@ -14,12 +14,13 @@ public class Tram extends Vehicle {
     /**
      * Tram constructor
      * Method creates new instance of Tram and creates a GUI object on canvas.
-     * @param position Coordinate of Trams within map
-     * @param speed Movement speed of given Tram
-     * @param onLine Line on which Tram transports people
+     *
+     * @param position   Coordinate of Trams within map
+     * @param speed      Movement speed of given Tram
+     * @param onLine     Line on which Tram transports people
      * @param identifier Unique Tram name
-     * @param street Dynamically changing street, on which Tram currently locates
-     * @param firstStop First stop on which Tram stops
+     * @param street     Dynamically changing street, on which Tram currently locates
+     * @param firstStop  First stop on which Tram stops
      * @param controller Main controller
      */
     public Tram(Coordinate position, double speed, Line onLine, String identifier, Street street, Stop firstStop, Controller controller, LocalTime time) {
@@ -29,10 +30,10 @@ public class Tram extends Vehicle {
         super.setStops();
 
         Polygon singleTram = new Polygon();
-        singleTram.getPoints().addAll((double)position.getX(), (double)position.getY() - 8,
-                (double)position.getX() + 8, (double)position.getY(),
-                (double)position.getX(), (double)position.getY() + 8,
-                (double)position.getX() - 8, (double)position.getY());
+        singleTram.getPoints().addAll(position.getX(), position.getY() - 8,
+                position.getX() + 8, position.getY(),
+                position.getX(), position.getY() + 8,
+                position.getX() - 8, position.getY());
         singleTram.setFill(Color.rgb(50, 50, 50));
         singleTram.setStroke(onLine.getColor());
         singleTram.setId(super.getName());
